final adjustments
