+++ e2e/src/lib.rs
@@ -74,38 +74,85 @@ mod tests {
         }
 
         // then
-        let state_submitting_finished = || async {
-            use services::{
-                block_bundler::port::Storage, block_importer::port::Storage as ImporterStorage,
-                state_committer::port::Storage as CommitterStorage,
-                state_listener::port::Storage as ListenerStorage,
-            };
-
-            let finished = stack
-                .db
-                .lowest_sequence_of_unbundled_blocks(0, 1)
-                .await?
-                .is_none()
-                && stack
-                    .db
-                    .oldest_nonfinalized_fragments(0, 1)
-                    .await?
-                    .is_empty()
-                && !stack.db.has_pending_txs().await?
-                && stack
-                    .db
-                    .missing_blocks(0, num_iterations * blocks_per_iteration)
-                    .await?
-                    .is_empty();
-
-            anyhow::Result::<_>::Ok(finished)
-        };
-
-        while !state_submitting_finished().await? {
+        while !state_submitting_finished(&stack.db, num_iterations * blocks_per_iteration).await? {
             tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
         }
 
         Ok(())
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn old_state_will_be_pruned() -> Result<()> {
+        use services::state_pruner::port::Storage;
+
+        // given
+        let show_logs = false;
+        let blob_support = true;
+        let stack = WholeStack::deploy_default(show_logs, blob_support).await?;
+
+        let num_iterations = 10;
+        let blocks_per_iteration = 100;
+
+        // when
+        for _ in 0..num_iterations {
+            let FuelNodeType::Local(node) = &stack.fuel_node else {
+                panic!("Expected local fuel node");
+            };
+
+            node.produce_transactions(100).await?;
+
+            let _ = stack
+                .fuel_node
+                .client()
+                .produce_blocks(blocks_per_iteration)
+                .await;
+        }
+
+        // then
+        while !state_submitting_finished(&stack.db, num_iterations * blocks_per_iteration).await? {
+            tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
+        }
+
+        let expected_table_sizes_before_pruning = services::state_pruner::port::TableSizes {
+            blob_transactions: 2,
+            fragments: 3,
+            transaction_fragments: 3,
+            bundles: 3,
+            blocks: 1300,
+            contract_transactions: 2,
+            contract_submisions: 2,
+        };
+        let table_sizes = stack.db.table_sizes().await?;
+        assert!(table_sizes >= expected_table_sizes_before_pruning);
+
+        // Sleep to make sure the pruner service had time to run
+        tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+
+        let table_sizes = stack.db.table_sizes().await?;
+        assert!(table_sizes < expected_table_sizes_before_pruning);
+
+        Ok(())
+    }
+
+    async fn state_submitting_finished(
+        db: &storage::DbWithProcess,
+        end_range: u32,
+    ) -> anyhow::Result<bool> {
+        use services::{
+            block_bundler::port::Storage, block_importer::port::Storage as ImporterStorage,
+            state_committer::port::Storage as CommitterStorage,
+            state_listener::port::Storage as ListenerStorage,
+        };
+
+        let finished = db
+            .lowest_sequence_of_unbundled_blocks(0, 1)
+            .await?
+            .is_none()
+            && db.oldest_nonfinalized_fragments(0, 1).await?.is_empty()
+            && !db.has_pending_txs().await?
+            && db.missing_blocks(0, end_range).await?.is_empty();
+
+        Ok(finished)
     }
 
     #[ignore = "meant for running manually and tweaking configuration parameters"]
