+++ e2e/src/whole_stack.rs
@@ -119,7 +119,9 @@ impl WholeStack {
                 .with_bundle_fragments_to_accumulate("3".to_owned())
                 .with_bundle_fragment_accumulation_timeout("10m".to_owned())
                 .with_new_bundle_check_interval("3s".to_owned())
-                .with_bundle_compression_level("level6".to_owned());
+                .with_bundle_compression_level("level6".to_owned())
+                .with_state_pruner_retention("14d".to_owned())
+                .with_state_pruner_run_interval("60s".to_owned());
 
             let committer = if blob_support {
                 committer_builder.with_blob_key_arn(secondary_key.id.clone())
@@ -228,7 +230,9 @@ async fn start_committer(
         .with_bundle_fragment_accumulation_timeout("5s".to_owned())
         .with_bundle_optimization_step("100".to_owned())
         .with_bundle_compression_level("level6".to_owned())
-        .with_new_bundle_check_interval("3s".to_owned());
+        .with_new_bundle_check_interval("3s".to_owned())
+        .with_state_pruner_retention("1s".to_owned())
+        .with_state_pruner_run_interval("30s".to_owned());
 
     let committer = if blob_support {
         committer_builder.with_blob_key_arn(secondary_key.id.clone())
